v0.8.0 (#46)
* Improved management of time when saving log.

Now it directly uses UNIX time without any timezone conversion.

* Re-ordered parameters of InventoriesQueries::addInventorySlotLog()

* Fixed inclusions/exclusions not working on MySQL + re-implemented support for meta when include/exclude blocks + fixed missing rollback binding parameter

* Fixed crash when using restricted block in UI + updated form library

* ConfigUpdater: now it maintains the previous settings where possible.

* Fixed user filter not working on MySQL

* Re-structured database tables (see description)

- Added constraint names for foreign keys
- Allow negative Y coordinate (ready for MC 1.17)
- Reduced X/Z coordinates to INTEGER instead of BIGINT
- All the column with data type BIGINT UNSIGNED are converted to SIGNED because PHP does not handle unsigned integer properly.

* Allowed radius in lookup command

* Fixed radius not optional in the lookup menu UI

* Forgot do add "ON DELETE CASCADE" in MySQL patch file

* Allow negative ID + allowed more space for meta (fixed #44)

* Fixed lookup taking the default radius instead of null

* Fixed #37

* Fixed entity filter not working

* Entities names are now saved with "#"

* Update Forms.php

* Fixed rollback radius not required but optional.

* Show "Just now" when an log just happend

* Show which options are implemented/changed/removed in the configuration when it's updated.

+++ src/matcracker/BedcoreProtect/storage/queries/Query.php
@@ -126,7 +126,7 @@ abstract class Query
         return yield Await::ONCE;
     }
 
-    final protected function addRawLog(string $uuid, Vector3 $position, string $worldName, Action $action, float $time): Generator
+    final protected function addRawLog(string $uuid, Vector3 $position, string $worldName, Action $action, int $time): Generator
     {
         return $this->executeInsert(QueriesConst::ADD_HISTORY_LOG, [
             'uuid' => mb_strtolower($uuid),
